Fixes thread affinity problem. (#70)
Any threads created in the messageTransceiver.init could inherit
the thread-affinity of the LoadTestRig-thread because the threadName
is set before the init method is called. And asynchronous, the
thread-affinity is set.

A concrete example is the cluster benchmarks that has the
client-driver-conductor running with the same isolated core as the
LoadTestRig and negatively impacting the latency tail.